Resolves issue #14. Also renamed Kaam24.py to kaam24.py.
